Using uint/int instead of uint32/int32 in ui package.

+++ ui/CheckBox.go
@@ -23,7 +23,7 @@ type CheckBox struct {
 //
 // Position and size will be adjusted to the current system DPI.
 func (me *CheckBox) Create(
-	parent Window, ctrlId, x, y int32, width, height uint32,
+	parent Window, ctrlId, x, y int, width, height uint,
 	text string, exStyles co.WS_EX, styles co.WS, btnStyles co.BS) *CheckBox {
 
 	x, y, width, height = _Util.MultiplyDpi(x, y, width, height)
@@ -39,7 +39,7 @@ func (me *CheckBox) Create(
 // Position will be adjusted to the current system DPI. The size will be
 // calculated to fit the text exactly.
 func (me *CheckBox) Create3State(
-	parent Window, ctrlId, x, y int32, text string) *CheckBox {
+	parent Window, ctrlId, x, y int, text string) *CheckBox {
 
 	return me.createAutoSize(parent, ctrlId, x, y, text, co.BS_AUTO3STATE)
 }
@@ -49,7 +49,7 @@ func (me *CheckBox) Create3State(
 // Position will be adjusted to the current system DPI. The size will be
 // calculated to fit the text exactly.
 func (me *CheckBox) Create2State(
-	parent Window, ctrlId, x, y int32, text string) *CheckBox {
+	parent Window, ctrlId, x, y int, text string) *CheckBox {
 
 	return me.createAutoSize(parent, ctrlId, x, y, text, co.BS_AUTOCHECKBOX)
 }
@@ -79,7 +79,7 @@ func (me *CheckBox) SetState(state co.BST) *CheckBox {
 // resizes the control to fit the text exactly.
 func (me *CheckBox) SetText(text string) *CheckBox {
 	cx, cy := me.calcIdealSize(me.Hwnd().GetParent(), text)
-	me.Hwnd().SetWindowPos(co.SWP_HWND_NONE, 0, 0, cx, cy,
+	me.Hwnd().SetWindowPos(co.SWP_HWND_NONE, 0, 0, uint32(cx), uint32(cy),
 		co.SWP_NOZORDER|co.SWP_NOMOVE)
 	me.Hwnd().SetWindowText(text)
 	return me
@@ -101,13 +101,13 @@ func (me *CheckBox) Text() string {
 }
 
 func (me *CheckBox) calcIdealSize(hReferenceDc win.HWND,
-	text string) (uint32, uint32) {
+	text string) (uint, uint) {
 
 	cx, cy := calcTextBoundBox(hReferenceDc, text, true)
-	cx += uint32(win.GetSystemMetrics(co.SM_CXMENUCHECK)) +
-		uint32(win.GetSystemMetrics(co.SM_CXEDGE)) // https://stackoverflow.com/a/1165052/6923555
+	cx += uint(win.GetSystemMetrics(co.SM_CXMENUCHECK)) +
+		uint(win.GetSystemMetrics(co.SM_CXEDGE)) // https://stackoverflow.com/a/1165052/6923555
 
-	cyCheck := uint32(win.GetSystemMetrics(co.SM_CYMENUCHECK))
+	cyCheck := uint(win.GetSystemMetrics(co.SM_CYMENUCHECK))
 	if cyCheck > cy {
 		cy = cyCheck // if the check is taller than the font, use its height
 	}
@@ -116,7 +116,7 @@ func (me *CheckBox) calcIdealSize(hReferenceDc win.HWND,
 }
 
 func (me *CheckBox) createAutoSize(
-	parent Window, ctrlId, x, y int32, text string, chbxStyles co.BS) *CheckBox {
+	parent Window, ctrlId, x, y int, text string, chbxStyles co.BS) *CheckBox {
 
 	x, y, _, _ = _Util.MultiplyDpi(x, y, 0, 0)
 	cx, cy := me.calcIdealSize(parent.Hwnd(), text)
